Fix implementation of a check
Before this commit, a null value will cause a wrong output of the evaluation
rule.

After this commit, we convert the input into an Option and then we are going to
check if the Option is a some or is a none.

Notice that this could change in future, depending on how the object is described
in the OpenAPI specification.